ui/gtk: keep the fraction when scaling relative motion
gd_motion_event() converts the pointer position from widget to guest
coordinates by dividing by gfx.scale_x/scale_y, and stores the result in
an int. In relative mode the guest is then sent the difference between
consecutive events, so the truncation is not a rounding detail: the
fraction is discarded on every event rather than accumulated.

With the window enlarged - zoom-to-fit on a display larger than the
guest, say - one host pixel is worth less than one guest pixel. Moving
the mouse slowly then produces a delta of zero every time and the guest
pointer does not move at all, while moving it quickly falls short of the
distance travelled, with the shortfall building up as the pointer moves.
Truncation is toward zero, so left and up lose the fraction differently
from right and down and the pointer drifts.

Keep last_x/last_y as doubles, report only the whole part and carry the
remainder into the next event.

+++ include/ui/gtk.h
@@ -129,8 +129,8 @@ struct GtkDisplayState {
     GtkWidget *notebook;
     int button_mask;
     gboolean last_set;
-    int last_x;
-    int last_y;
+    double last_x;
+    double last_y;
     int grab_x_root;
     int grab_y_root;
     VirtualConsole *kbd_owner;
